Try reloading once if no GeoIP cookie found
If server-side code for Special:FundraiserRedirector doesn't find a GeoIP cookie,
reload the page only once. In cases where no GeoIP cookie was previously available
for this domain, but has been set on this request by the front-end cache, reloading
should make it available to server-side code.

See the Phabricator task linked below for the  rationale for this approach to fixing
this issue, and for testing instructions.

Bug: T317427
Change-Id: If008d5d0d0ea9ad3ee4284ab240149bd1587f94c